config.py
import os
from pathlib import Path
from dotenv import load_dotenv

load_dotenv()

# Tokens (nom nom nom, tasty 🤤😝)
TOKEN = os.getenv("DISCORD_TOKEN")
LOGGING_DEBUG_MODE = os.getenv("LOGGING_DEBUG_MODE", False)
if not TOKEN:
    raise SystemExit("Set DISCORD_TOKEN in .env")

# Bot settings
COMMAND_PREFIX = "!!"
