requirements.txt
discord.py==2.6.4
dopamine-framework==1.5.0
python-dotenv==1.2.1